refactor: split allowance statistics
